+++ src/server/js/functions.js
@@ -16,5 +16,28 @@ Roles.server = {
             })
         }
         return ret;
+    },
+
+    // remove all roles for the user
+    //  returns true|false
+    removeAllRolesFromUser( user ){
+        let ret = -1;
+        if( user ){
+            let id = null;
+            if( _.isString( user )){
+                id = user;
+            } else if( _.isObject( user ) && user._id ){
+                id = user._id;
+            }
+            if( id ){
+                ret = Meteor.roleAssignment.remove({ 'user._id': id });
+            } else {
+                console.warn( 'removeAllRolesFromUser() unable to find an identifier', user );
+            }
+        } else {
+            console.warn( 'removeAllRolesFromUser() user is falsy', user );
+        }
+        console.debug( 'removeAllRolesFromUser() ret=', ret )
+        return ret;
     }
 };